Sync v2_9_r0_3 from private repo
- __version__.py as single source of truth for the codebase version;
  entry script exposes it via --version and logs it on startup.
- parse_bool() helper for strict boolean coercion, used for ai_enabled
  and the concatenate trim parameter.
- config_loader.py: date_validation values now accept native YAML date
  objects, not just strings; non-numeric numeric_validation bounds
  reported as a config error instead of crashing.
- excel_utils.append_row_safe() guards against formula injection in
  generated worksheet rows; applied in Importer and IssuesWriter.
- Failed-enrichment dependency skip in validation_engine.py now scoped
  per (SheetName, ColumnName) instead of a bare column name.
- counter4group() returns int64 group counters instead of float.
- CHANGELOG.md and etc/versioning_scheme.md added, documenting the
  vN_L_rK_M versioning scheme this release follows.
- README.md: updated Usage section and FS link.

Full suite: 488/488 passing.

+++ src/v2/functions/builtin/counter4group.py
@@ -55,4 +55,4 @@ def counter4group(context: dict) -> list:
     # Restore original row order
     working = working.sort_values("_original_index")
 
-    return working["_group_counter"].tolist()
+    return working["_group_counter"].astype("int64").tolist()
